created compnents Chats and Input
